+++ source/lambda/deploy_es/index.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: MIT-0
 __copyright__ = ('Copyright Amazon.com, Inc. or its affiliates. '
                  'All Rights Reserved.')
-__version__ = '2.10.3'
+__version__ = '2.10.4'
 __license__ = 'MIT-0'
 __author__ = 'Akihiro Nakajima'
 __url__ = 'https://github.com/aws-samples/siem-on-amazon-opensearch-service'
@@ -158,7 +158,7 @@ access_policies_json = json.dumps(access_policies)
 
 config_domain = {
     'DomainName': AOS_DOMAIN,
-    'EngineVersion': 'OpenSearch_2.11',
+    'EngineVersion': 'OpenSearch_2.19',
     'ClusterConfig': {
         'InstanceType': 't3.medium.search',
         'InstanceCount': 1,